backend/requirements.txt
# FastAPI and Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlmodel>=0.0.16
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.25
greenlet>=3.0.0

# Migrations
alembic>=1.13.1

# Configuration
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Authentication/JWT
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Email validation
email-validator>=2.1.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
aiosqlite>=0.19.0

# Development
python-multipart>=0.0.7
